repo changes

+++ ansible/s3motr-build-depencies.sh
@@ -23,26 +23,15 @@ set -x
 SCRIPT_PATH=$(readlink -f "$0")
 SCRIPT_DIR=$(dirname "$SCRIPT_PATH")
 
-#motr_script_path=$SCRIPT_DIR/../third_party/motr/scripts
-#if [[ -d "$motr_script_path" && -f "$motr_script_path/install-build-deps" ]];
-#then
-#    sh $motr_script_path/install-build-deps
-#else
+motr_script_path=$SCRIPT_DIR/../third_party/motr/scripts
+if [[ -d "$motr_script_path" && -f "$motr_script_path/install-build-deps" ]];
+then
+    sh $motr_script_path/install-build-deps
+else
     # third_party submodules are not loaded yet hence do fresh clone and
     # run install-build-deps script.
     # This needs GitHub id and Personal Access Token
-#    git clone https://github.com/Seagate/cortx-motr.git
-#    ./cortx-motr/scripts/install-build-deps
-#    rm -rf cortx-motr
-#fi
-
-git clone https://github.com/Seagate/cortx-motr.git
-cd ./cortx-motr
-# checkout speicif code for Kernel & luster version changes
-git checkout 4b3f850bc971a9570190c8b8b5c4d26bdf3c4790
-
-sh scripts/install-build-deps
-cd ..
-
-rm -rf cortx-motr
-
+    git clone https://github.com/Seagate/cortx-motr.git
+    ./cortx-motr/scripts/install-build-deps
+    rm -rf cortx-motr
+fi
